+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Year electricity widget with intersections observer (lazy loading).
+- Lazy loading month electricity widget (intersections observer).
+
+### Fixed
+- Improvements header modal.
+
 ## [1.14.2] - 2021-11-14
 ### Fixed
 - Removed Year widget because of performance issues.